Fix token reuse issue.
Token where being used for two purposes, identifying a connection and
as an index to the handler on the slab.  The problem is that when a
connection is closed this index is freed on the slab and given to the
next connection.  This opens up a situations where we can easily send
messages to the wrong client.  Steps to reproduce this problem are as
follows:

1. Client A connects.
2. The server initiates some async operation and holds a clone of
	client A's ws::Sender to send the results.
3. Client A disconnects freeing its token.
4. Client B connects
5. The server finishes its async operation and send the results using
	the ws::Sender it thinks belongs to A.
6. Client B receives the message.

The solution is to give each connection a unique connection ID in
addition to the token.  The token is used to quickly get the socket
from the slab and the connection ID is used to insure this is the socket
we think it is.

Signed-off-by: Jason Housley <HousleyJK@gmail.com>